Move FedCoRe tests into central suite
